Avoid using of sh builtin echo command.
On Mac OS X, the /bin/sh shell has a builtin echo command that does not recognize the -n option for linefeed supression. To keep the test script portable, the script now uses the echo command installed on the machine (usually in /bin/echo). The which function is used to make sure that the echo command is correctly located.