Add support for writing of XML observation definition files.
In case that the observation container holds more than a single observation, ctobssim now outputs an observation definition XML file in addition to the event files. The filenames of the event files will be stored in the observation definition XML file so that subsequent ctools can handle all observational data together.