Improve ctlike output
ctlike now computes the observed number of events and dumps the information just before the predicted number of events into the log file. The difference between the observed and predicted number of events is also given so that one knows immediately if the fit behaved as it should.

This commit closes actions #87 and #88 and the feature #54.