Add OPENMP_LIB_LDFLAGS for OpenMP support on Mac OS X.
For OpenMP to work properly on Mac OS X, the -lgomp library has to be linked. For this purpose, a new variable OPENMP_LIB_LDFLAGS has been created that is set by configure to -lgomp in the presence of OpenMP support.

The ctobssim Makefile.am has been modified to take this parameter into account during linking.