Make compliant to MacPorts gcc compiler.
A Darwin kluge consisted of linking statically with the ctools library for unit testing. It turned out that this is no longer necessary, so I removed the Darwin kluge of static linking.

Furthermore, a post-processing step was added that creates a symbolic link to the directory that contains the shared wrapper library. In that way, we do not need a double compilation anymore (before, we compiled once for the install and a second time for the unit test). This should now speed up the unit testing step.