Correct test if library can be dlopen()ed.
The test for dlopen()ing libgomp.so.1 failed because for some reason the test code did not find the GammaLib shared library. I'm not sure why this happened.

To make the test work, the LIBS variable is now set explicitely to the library we need for the test (-ldl), excluding the GammaLib shared library. This allows the test to suceed.