Add usepnt parameter and adapt to new GTime interface.
The usepnt parameter now signals when the pointing should be used as ROI centre. This allows arbitrary values for ra/dec (before, ra=0 and dec=0 signalled that the pointing should be used).

The tool now uses the new GTime interface. In particular, the input time is now given in CTA reference time in seconds.