Adapt to new GCaldb interface.
A new logic has been implemented concerning the caldb parameter. If the name specified by caldb is a directory, the directory is set as the root directory of the calibration database. The file specified by irf must then be a file provided within this root directory.

If the name is not a directory it is used as the mission name that is used to build the calibration directory path. The irf is then the response name.