Correct sim function, and add set_obs and set_obs_list functions.
The CTA observation container was not copied in the sim function following yesterday's merge, this was corrected.

The set function is now obsolete, and has been renamed to set_obs for clarity. The set function is still there, but a warning is issued if it is still used.

A set_obs_list function has been added that allows setting lists of observations.