Introduce cscript Python class from which all Python cscripts are derived.
We cannot use the ctool base class for the Python scripts as this is an abstract class an Python does not know the concept of abstract classes. Hence for Python we need a real class to derive from, which we implement by adding the cscript class to the ctool.i interface.