Correct Python test script.
There was an extract blank following a \ character that led to a failure, but at the same time, the failure was not catched since the script checked the presence of a file that was generated in a test before. I renamed the file that is created to a unique name, and remov of coarse the extra space.