Added warning and updated test cases for ctphase (#1982)
Added a warning in ctphase that is written into the log file when either the first or second derivative term in the phase computation is above 1e8, as at this point precision issues can manifest in the computed phase values.
Updated test cases and fixed some bugs that allow the tests of ctphase to pass.