Added reading phase info from XML file (#1982)
Added to ctphase the capability to read the phase information from the temporal component of a source in an XML file. User needs to specify the name of the model XML file and the name of the source. If either of these is missing, then ctphase defaults to reading the phase information from the ctphase.par file. If both model filename and source are defined, checks are performed to make sure the values are reasonable.