Updated model trimming and parallelization in ctmodel (#2189)
Updated the model trimming so that the models that are outside an observation are offloaded to another model container and then re-appended after the evaluation.
Parallelization scheduling set to "static" with a chunk size of 1. This will slightly help with the issue in which multiple observations in a row do not contribute to the model cube and thus dividing the observations into "n-observations/n-cores" is not ideal.