Refactor test source position and pointing direction setting in cssens (#3555).
A _set_source() method was introduced in cssens that deals with the setting of the test source direction independent of the code in the _set_obs() method, since that latter method is not necessarily called. The _set_obs() method now only deals with setting of the pointing direction of a single observation, in case that no observation definition XML file was specified on input.